+++ AHHA.Domain/Entities/Accounts/AR/ArDocSetOffDt.cs
@@ -8,15 +8,15 @@ namespace AHHA.Core.Entities.Accounts.AR
 {
     public class ArDocSetOffDt
     {
-        public byte CompanyId { get; set; }
+        public Int16 CompanyId { get; set; }
         public long SetoffId { get; set; }
         public string SetoffNo { get; set; }
         public string ReferenceNo { get; set; }
-        public byte ItemNo { get; set; }
-        public byte TransactionId { get; set; }
+        public Int16 ItemNo { get; set; }
+        public Int16 TransactionId { get; set; }
         public long DocumentId { get; set; }
         public string DocumentNo { get; set; }
-        public byte DocCurrencyId { get; set; }
+        public Int16 DocCurrencyId { get; set; }
         public decimal DocExhRate { get; set; }
         public DateTime DocAccountDate { get; set; }
         public DateTime DocDueDate { get; set; }
