+++ AHHA.Domain/Entities/Accounts/AR/ArOutstandTransactions.cs
@@ -8,15 +8,15 @@ namespace AHHA.Core.Entities.Accounts.AR
 {
     public class ArOutstandTransactions
     {
-        public byte CompnyId { get; set; }
-        public byte TransactionId { get; set; }
+        public Int16 CompnyId { get; set; }
+        public Int16 TransactionId { get; set; }
         public long DocumentId { get; set; }
         public string DocumentNo { get; set; }
         public string ReferenceNo { get; set; }
         public DateTime AccountDate { get; set; }
         public DateTime DueDate { get; set; }
         public int CustomerId { get; set; }
-        public byte CurrencyId { get; set; }
+        public Int16 CurrencyId { get; set; }
         public decimal ExhRate { get; set; }
         public decimal TotAmt { get; set; }
         public decimal TotLocalAmt { get; set; }
