+++ AHHA.Domain/Entities/Accounts/AR/ArReceiptHd.cs
@@ -8,22 +8,22 @@ namespace AHHA.Core.Entities.Accounts.AR
 {
     public class ArReceiptHd
     {
-        public byte CompanyId { get; set; }
+        public Int16 CompanyId { get; set; }
         public long ReceiptId { get; set; }
         public string ReceiptNo { get; set; }
         public string ReferenceNo { get; set; }
         public DateTime TrnDate { get; set; }
         public DateTime AccountDate { get; set; }
-        public short BankId { get; set; }
-        public short PaymentTypeId { get; set; }
+        public Int32 BankId { get; set; }
+        public Int32 PaymentTypeId { get; set; }
         public string ChequeNo { get; set; }
         public DateTime ChequeDate { get; set; }
         public int CustomerId { get; set; }
-        public byte CurrencyId { get; set; }
+        public Int16 CurrencyId { get; set; }
         public decimal ExhRate { get; set; }
         public decimal TotAmt { get; set; }
         public decimal TotLocalAmt { get; set; }
-        public byte RecCurrencyId { get; set; }
+        public Int16 RecCurrencyId { get; set; }
         public decimal RecExhRate { get; set; }
         public decimal RecTotAmt { get; set; }
         public decimal RecTotLocalAmt { get; set; }
