+++ AHHA.Domain/Entities/Accounts/CB/CBPettyCashHd.cs
@@ -8,17 +8,17 @@ namespace AHHA.Core.Entities.Accounts.CB
 {
     public class CBPettyCashHd
     {
-        public byte CompanyId { get; set; }
+        public Int16 CompanyId { get; set; }
         public long ReceiptId { get; set; }
         public string ReceiptNo { get; set; }
         public string ReferenceNo { get; set; }
         public DateTime TrnDate { get; set; }
         public DateTime AccountDate { get; set; }
-        public short BankId { get; set; }
-        public byte CurrencyId { get; set; }
+        public Int32 BankId { get; set; }
+        public Int16 CurrencyId { get; set; }
         public decimal ExhRate { get; set; }
         public decimal CtyExhRate { get; set; }
-        public short PaymentTypeId { get; set; }
+        public Int32 PaymentTypeId { get; set; }
         public string ChequeNo { get; set; }
         public DateTime ChequeDate { get; set; }
         public decimal BankChgAmt { get; set; }
@@ -36,12 +36,12 @@ namespace AHHA.Core.Entities.Accounts.CB
         public string Remarks { get; set; }
         public string PayeeTo { get; set; }
         public string ModuleFrom { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
         public bool IsCancel { get; set; }
-        public short CancelById { get; set; }
+        public Int32 CancelById { get; set; }
         public DateTime CancelDate { get; set; }
         public string CancelRemarks { get; set; }
     }
