+++ AHHA.Domain/Entities/Accounts/GL/GLBalances_GLCode.cs
@@ -8,11 +8,11 @@ namespace AHHA.Core.Entities.Accounts.GL
 {
     public class GLBalances_GLCode
     {
-        public byte CompanyId { get; set; }
-        public byte CurrencyId { get; set; }
-        public short FinYear { get; set; }
-        public byte FinMonth { get; set; }
-        public short GLId { get; set; }
+        public Int16 CompanyId { get; set; }
+        public Int16 CurrencyId { get; set; }
+        public Int32 FinYear { get; set; }
+        public Int16 FinMonth { get; set; }
+        public Int32 GLId { get; set; }
         public bool IsDebit { get; set; }
         public decimal TotAmt { get; set; }
         public decimal TotLocalAmt { get; set; }
