+++ AHHA.Domain/Entities/Accounts/GL/GLPeriodClose.cs
@@ -8,26 +8,26 @@ namespace AHHA.Core.Entities.Accounts.GL
 {
     public class GLPeriodClose
     {
-        public byte CompnyId { get; set; }
-        public short FinYear { get; set; }
-        public byte FinMonth { get; set; }
+        public Int16 CompnyId { get; set; }
+        public Int32 FinYear { get; set; }
+        public Int16 FinMonth { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public bool IsArClose { get; set; }
-        public short ArCloseById { get; set; }
+        public Int32 ArCloseById { get; set; }
         public DateTime ArCloseDate { get; set; }
         public bool IsApClose { get; set; }
-        public short ApCloseById { get; set; }
+        public Int32 ApCloseById { get; set; }
         public DateTime ApCloseDate { get; set; }
         public bool IsCbClose { get; set; }
-        public short CbCloseById { get; set; }
+        public Int32 CbCloseById { get; set; }
         public DateTime CbCloseDate { get; set; }
         public bool IsGlClose { get; set; }
-        public short GlCloseById { get; set; }
+        public Int32 GlCloseById { get; set; }
         public DateTime GlCloseDate { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
