+++ AHHA.Domain/Entities/Admin/AccountType.cs
@@ -10,17 +10,17 @@ namespace AHHA.Core.Entities.Admin
     public class AccountType
     {
         [Key]
-        public byte AccTypeId { get; set; }
-        public byte CompanyId { get; set; }
+        public Int16 AccTypeId { get; set; }
+        public Int16 CompanyId { get; set; }
         public string AccTypeCode { get; set; }
         public string AccTypeName { get; set; }
-        public byte SeqNo { get; set; }
+        public Int16 SeqNo { get; set; }
         public string AccGroupName { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
