+++ AHHA.Domain/Entities/Admin/AdmErrorLog.cs
@@ -11,15 +11,15 @@ namespace AHHA.Core.Entities.Admin
     {
         [Key]
         public long ErrId { get; set; }
-        public byte CompanyId { get; set; }
-        public byte ModuleId { get; set; }
-        public short TransactionId { get; set; }
+        public Int16 CompanyId { get; set; }
+        public Int16 ModuleId { get; set; }
+        public Int32 TransactionId { get; set; }
         public long DocumentId { get; set; }
         public string DocumentNo { get; set; }
         public string TblName { get; set; }
-        public byte ModeId { get; set; }
+        public Int16 ModeId { get; set; }
         public string Remarks { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         //public DateTime CreateDate { get; set; }
     }
 }
