+++ AHHA.Domain/Entities/Admin/Company.cs
@@ -10,7 +10,7 @@ namespace AHHA.Core.Entities.Admin
     public class Company
     {
         [Key]
-        public byte CompanyId { get; set; }
+        public Int16 CompanyId { get; set; }
         public string CompanyCode { get; set; }
         public string CompanyName { get; set; }
         public string RegistrationNo { get; set; }
@@ -18,9 +18,9 @@ namespace AHHA.Core.Entities.Admin
         public long AddressId { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
