+++ AHHA.Domain/Entities/Admin/Module.cs
@@ -10,10 +10,10 @@ namespace AHHA.Core.Entities.Admin
     public class Module
     {
         [Key]
-        public byte ModuleId { get; set; }
+        public Int16 ModuleId { get; set; }
         public string ModuleCode { get; set; }
         public string ModuleName { get; set; }
-        public byte SeqNo { get; set; }
+        public Int16 SeqNo { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
         public string CreateBy { get; set; }
