+++ AHHA.Domain/Entities/Admin/ShareData.cs
@@ -5,11 +5,11 @@ namespace AHHA.Core.Entities.Admin
     [Keyless]
     public class ShareData
     {
-        public byte ModuleId { get; set; }
-        public short TransactionId { get; set; }
+        public Int16 ModuleId { get; set; }
+        public Int32 TransactionId { get; set; }
         public bool ShareToAll { get; set; }
-        public byte CompanyId { get; set; }
-        public byte SetId { get; set; }
+        public Int16 CompanyId { get; set; }
+        public Int16 SetId { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
         public string CreateBy { get; set; }
