+++ AHHA.Domain/Entities/Admin/User.cs
@@ -10,17 +10,17 @@ namespace AHHA.Core.Entities.Admin
     public class User
     {
         [Key]
-        public short UserId { get; set; }
+        public Int32 UserId { get; set; }
         public string UserCode { get; set; }
         public string UserName { get; set; }
         public string UserPassword { get; set; }
         public string UserEmail { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
-        public short UserGroupId { get; set; }
+        public Int32 UserGroupId { get; set; }
     }
 }
