+++ AHHA.Domain/Entities/Admin/UserGroup.cs
@@ -10,14 +10,14 @@ namespace AHHA.Core.Entities.Admin
     public class UserGroup
     {
         [Key]
-        public short UserGroupId { get; set; }
+        public Int32 UserGroupId { get; set; }
         public string UserGroupCode { get; set; }
         public string UserGroupName { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
