+++ AHHA.Domain/Entities/Admin/UserGroupRights.cs
@@ -12,17 +12,17 @@ namespace AHHA.Core.Entities.Admin
     public class UserGroupRights
     {
         [Key]
-        public short UserGroupId { get; set; }
+        public Int32 UserGroupId { get; set; }
         [Key]
-        public byte ModuleId { get; set; }
+        public Int16 ModuleId { get; set; }
         [Key]
-        public short TransactionId { get; set; }
+        public Int32 TransactionId { get; set; }
         public bool IsRead { get; set; }
         public bool IsEdit { get; set; }
         public bool IsDelete { get; set; }
         public bool IsExport { get; set; }
         public bool IsPrint { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
     }
 }
