+++ AHHA.Domain/Entities/Masters/M_AccountSetup.cs
@@ -10,16 +10,16 @@ namespace AHHA.Core.Entities.Masters
     public class M_AccountSetup
     {
         [Key]
-        public byte AccSetupId { get; set; }
-        public byte CompanyId { get; set; }
+        public Int16 AccSetupId { get; set; }
+        public Int16 CompanyId { get; set; }
         public string AccSetupCode { get; set; }
         public string AccSetupName { get; set; }
-        public byte AccSetupCategoryId { get; set; }
+        public Int16 AccSetupCategoryId { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
