+++ AHHA.Domain/Entities/Masters/M_AccountSetupCategory.cs
@@ -10,14 +10,14 @@ namespace AHHA.Core.Entities.Masters
     public class M_AccountSetupCategory
     {
         [Key]
-        public byte AccSetupCategoryId { get; set; }
+        public Int16 AccSetupCategoryId { get; set; }
         public string AccSetupCategoryCode { get; set; }
         public string AccSetupCategoryName { get; set; }
         public string Remarks { get; set; }
         public bool IsActive { get; set; }
-        public short CreateById { get; set; }
+        public Int32 CreateById { get; set; }
         public DateTime CreateDate { get; set; }
-        public short EditById { get; set; }
+        public Int32 EditById { get; set; }
         public DateTime EditDate { get; set; }
     }
 }
